+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project follows versions of format `{year}.{month}.{patch_number}`.
 
 ## [Unreleased]
 
+## [2026.06.0-rc.1] - 2026-05-15
+
 ### Changed
 
 - [BREAKING] Changed the output of `read_results` to return a stronger-typed `CalibrationResults`